Producthunt_RapidAPI/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-multipart
python-dotenv